Factor out common server tunables.
