+++ src/types/configuration/CentralSystemFrontEndConfiguration.ts
@@ -1,9 +1,6 @@
-import { ServerProtocol } from '../Server';
+import CentralSystemServerConfiguration from './CentralSystemServerConfiguration';
 
-export default interface CentralSystemFrontEndConfiguration {
-  protocol: ServerProtocol;
-  host: string;
-  port: number;
+export default interface CentralSystemFrontEndConfiguration extends CentralSystemServerConfiguration {
   distEnabled?: boolean;
   distPath?: string;
 }
