+++ src/types/configuration/CentralSystemServerConfiguration.ts
